Maringo_club/requirements1.txt
Flask>=2.3
Flask-RESTful>=0.3
Flask-CORS>=4.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-WTF>=1.2
python-dotenv>=1.0
pandas>=2.2
matplotlib>=3.8
reportlab>=4.0
openpyxl>=3.1
Pillow>=10.0
email-validator>=2.1
